Add basic search handler.
Only supports substring search for now. If a search has only one result,
it forwards on to the entry for that page; otherwise, it displays the
results (or UI telling the user there are no results). But the
forwarding does not happen until the user is authenticated, so no
information is leaked to unauthenticated users.