Add editing/creating/deleting entries.
One issue is that once edit mode is activated, the auto-hide is disabled
permanently. It's not so clear what the solution is: if we auto-hide
content that is being edited, a password might be lost before it is
committed.

The code could be cleaner, but the functionality works, so this is a
good stopping point.