Add `key` package.
This will allow more encryption methods to be added, and make it easy to
e.g. write a utility to switch between different key types. So far, the
PGP/pass-based store is the only one implemented, but this commit is the
last thing I need to do before I can implement a new encryption method.