Add native Harpocrates keys.
No more PGP! There are lots of TODOs in the code, but this is a good
stopping point -- harpd_debug now serves based on native-Harpocrates
encryption by default, and things seem to be working.

Native Harpocrates encryption uses AES-GCM to encrypt & authenticate
entry content on disk. The encryption key (EK) is generated randomly,
and encrypted on disk with a key-encryption key (KEK) that is generated
from the user's passphrase via scrypt.