Alert when a non-fully-authed session is closed for any reason.
Previously, an alert would fire if a session timed out after a password
was given but before U2F authentication was completed. This could
indicate that the password has leaked or been guessed. However, the
/logout endpoint would bypass this alert as of commit e17f7c5. Instead,
now just fire an alert any time a non-fully-authenticated session is
closed for any reason, to avoid reintroducing similar alert-bypass bugs
in the future.