Tear out MFA device counter value tracking.
Tracking counter values is not necessary for the security of the U2F or
(soon) WebAuthn challenge/response protocol. Counter-value tracking was
added to the U2F standard with the intent to detect cloned MFA devices,
and then I suppose cargo-culted into the WebAuthn standard.  Since MFA
device cloning is not in my threat model, I remove this as unnecessary
complexity.

[Discussion about why the counter is not useful:
https://github.com/Yubico/php-u2flib-server/issues/5]