Allow refreshes after editing password content.
Previously, after editing content, the POST request itself would return
the updated content HTML. This meant that attempting to refresh the page
(for example, to refresh the content-hiding timer) would result in the
browser asking if the user wanted to resubmit the form. Now, the POST
request will always result in a redirect to a GET request for the same
URI--from the user's POV, they see the same behavior, except now they
can refresh the page without being harassed by the browser (or actually
editing the password content again).