New Password Generator UI.
The user can now control the length & character set used for generated
passwords, as well as rerolling passwords & copying them directly to the
clipboard. I decided to generate passwords in the client--I am wary of
Javascript, but I found a way to implement the necessary debiased random
number generation in a simple enough way that I am fairly confident that
I got it right.

I also removed a previous implementation of the debiased random number
generation in Go, since it's no longer necessary.